+++ CHANGELOG.md
@@ -2927,3 +2927,4 @@
 - [2026-07-17 21:04] feat(forms): add async forms stream provider
 - [2026-07-17 22:21] docs(cache): update readme and API contracts for cache
 - [2026-07-17 22:57] fix(storage): fix memory leak by cancelling storage subscription
+- [2026-07-18 09:00] feat(cache): implement shimmer loading skeleton in cache